requirements.txt
Django==4.1.5
Pillow==9.4.0
django-environ==0.9.0
